Glitches with the conversion were discovered when investigating issue #340.
Sequences of strings are turned into numpy arrays of dtype "O", and these
where converted to R list. However, this can be seen as an incorrect
or unwanted conversion when all elements in the array are strings.

Converting to R string vectors (rather than R lists) seems more natural,
and this is what the fix is now doing, but it is also creating a situation
where the type of the output depends on the exact nature of the elements
in the input vectors. Might work for interative work, but in other situations
it will mean that the type of the output might have to be tested after
conversion.